Changed all SQL to use parameters.

+++ Netus2_DatabaseConnection/daoImplementations/JctCourseGradeDaoImpl.cs
@@ -4,6 +4,7 @@ using Netus2_DatabaseConnection.utilityTools;
 using System;
 using System.Collections.Generic;
 using System.Data;
+using System.Data.SqlClient;
 using System.Text;
 
 namespace Netus2_DatabaseConnection.daoImplementations
@@ -12,58 +13,86 @@ namespace Netus2_DatabaseConnection.daoImplementations
     {
         public void Delete(int courseId, int gradeId, IConnectable connection)
         {
-            StringBuilder sql = new StringBuilder("DELETE FROM jct_course_grade WHERE 1=1 ");
-            sql.Append("AND course_id = " + courseId + " ");
-            sql.Append("AND enum_grade_id = " + gradeId);
+            if (courseId <= 0 || gradeId < 0)
+                throw new Exception("Cannot delete a record from jct_course_grade " +
+                    "without a database-assigned ID for both courseId and gradeId." +
+                    "\ncourseId: " + courseId +
+                    "\ngradeId: " + gradeId);
 
-            connection.ExecuteNonQuery(sql.ToString());
+            StringBuilder sql = new StringBuilder("DELETE FROM jct_course_grade WHERE 1=1 ");
+            sql.Append("AND course_id = @course_id ");
+            sql.Append("AND enum_grade_id = @enum_grade_id");
+
+            List<SqlParameter> parameters = new List<SqlParameter>();
+            parameters.Add(new SqlParameter("@course_id", courseId));
+            parameters.Add(new SqlParameter("@enum_grade_id", gradeId));
+
+            connection.ExecuteNonQuery(sql.ToString(), parameters);
         }
 
         public DataRow Read(int courseId, int gradeId, IConnectable connection)
         {
-            StringBuilder sql = new StringBuilder("SELECT * FROM jct_course_grade WHERE 1=1 ");
-            sql.Append("AND course_id = " + courseId + " ");
-            sql.Append("AND enum_grade_id = " + gradeId);
+            string sql = "SELECT * FROM jct_course_grade WHERE 1=1 " +
+            "AND course_id = @course_id " +
+            "AND enum_grade_id = @enum_grade_id";
 
-            List<DataRow> results = Read(sql.ToString(), connection);
-            if (results.Count == 1)
-                return results[0];
-            else if (results.Count == 0)
+            List<SqlParameter> parameters = new List<SqlParameter>();
+            parameters.Add(new SqlParameter("@course_id", courseId));
+            parameters.Add(new SqlParameter("@enum_grade_id", gradeId));
+
+            List<DataRow> results = Read(sql, connection, parameters);
+
+            if (results.Count == 0)
                 return null;
-            else
-                throw new Exception("The jct_course_grade table contains a duplicate record.\n" +
-                    "course_id = " + courseId + ", grade_id = " + gradeId);
+            else 
+                return results[0];
         }
 
-        public List<DataRow> Read(int courseId, IConnectable connection)
+        public List<DataRow> Read_AllWithCourseId(int courseId, IConnectable connection)
         {
-            StringBuilder sql = new StringBuilder("SELECT * FROM jct_course_grade WHERE 1=1 ");
-            sql.Append("AND course_id = " + courseId);
+            string sql = "SELECT * FROM jct_course_grade WHERE " +
+            "course_id = @course_id";
 
-            return Read(sql.ToString(), connection);
+            List<SqlParameter> parameters = new List<SqlParameter>();
+            parameters.Add(new SqlParameter("@course_id", courseId));
+
+            return Read(sql, connection, parameters);
         }
 
-        private List<DataRow> Read(string sql, IConnectable connection)
+        public List<DataRow> Read_AllWithGradeId(int gradeId, IConnectable connection)
+        {
+            string sql = "SELECT * FROM jct_course_grade WHERE " +
+            "enum_grade_id = @enum_grade_id";
+
+            List<SqlParameter> parameters = new List<SqlParameter>();
+            parameters.Add(new SqlParameter("@enum_grade_id", gradeId));
+
+            return Read(sql, connection, parameters);
+        }
+
+        private List<DataRow> Read(string sql, IConnectable connection, List<SqlParameter> parameters)
         {
             DataTable dtJctCourseGrade = DataTableFactory.CreateDataTable_Netus2_JctCourseGrade();
-            dtJctCourseGrade = connection.ReadIntoDataTable(sql, dtJctCourseGrade);
+            dtJctCourseGrade = connection.ReadIntoDataTable(sql, dtJctCourseGrade, parameters);
 
             List<DataRow> jctCourseGradeDaos = new List<DataRow>();
             foreach(DataRow row in dtJctCourseGrade.Rows)
-            {
                 jctCourseGradeDaos.Add(row);
-            }
 
             return jctCourseGradeDaos;
         }
 
         public DataRow Write(int courseId, int gradeId, IConnectable connection)
         {
-            StringBuilder sql = new StringBuilder("INSERT INTO jct_course_grade (course_id, enum_grade_id) VALUES (");
-            sql.Append(courseId + ", ");
-            sql.Append(gradeId + ")");
+            string sql = "INSERT INTO jct_course_grade (" +
+                "course_id, enum_grade_id) VALUES (" +
+                "@course_id, @enum_grade_id)";
 
-            connection.ExecuteNonQuery(sql.ToString());
+            List<SqlParameter> parameters = new List<SqlParameter>();
+            parameters.Add(new SqlParameter("@course_id", courseId));
+            parameters.Add(new SqlParameter("@enum_grade_id", gradeId));
+
+            connection.ExecuteNonQuery(sql, parameters);
 
             DataRow jctCourseGradeDao = DataTableFactory.CreateDataTable_Netus2_JctCourseGrade().NewRow();
             jctCourseGradeDao["course_id"] = courseId;
